# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/members": {
      "get": {
        "operationId": "AdminMemberController_findAll_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by member status",
            "schema": {
              "enum": ["PENDING", "APPROVED", "REJECTED"],
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filter by active/expired status",
            "schema": {
              "enum": ["ACTIVE", "EXPIRED"],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name, email, phone, or memberId",
            "schema": {
              "example": "John",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (for pagination)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (for pagination)",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all members with filters",
        "tags": ["Admin - Members"]
      }
    },
    "/api/v1/members/{id}": {
      "get": {
        "operationId": "AdminMemberController_getMemberDetail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get member detail (membership info + images)",
        "tags": ["Admin - Members"]
      },
      "delete": {
        "operationId": "AdminMemberController_deleteMember_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete member (deletedAt)",
        "tags": ["Admin - Members"]
      }
    },
    "/api/v1/members/{id}/approve": {
      "patch": {
        "operationId": "AdminMemberController_approveMember_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve member registration (sets status = APPROVED, calculates startDate & endDate)",
        "tags": ["Admin - Members"]
      }
    },
    "/api/v1/members/{id}/reject": {
      "patch": {
        "operationId": "AdminMemberController_rejectMember_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject member registration (status = REJECTED)",
        "tags": ["Admin - Members"]
      }
    },
    "/api/v1/members/{id}/restore": {
      "patch": {
        "operationId": "AdminMemberController_restoreMember_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore soft-deleted member",
        "tags": ["Admin - Members"]
      }
    },
...
    "/api/v1/checkins": {
      "get": {
        "operationId": "CheckInController_findAll_v1",
        "parameters": [
          {
            "name": "memberId",
            "required": false,
            "in": "query",
            "description": "Filter by member ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by check-in status",
            "schema": {
              "enum": ["ALLOWED", "DENIED"],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for date range filter (ISO 8601)",
            "schema": {
              "example": "2024-01-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for date range filter (ISO 8601)",
            "schema": {
              "example": "2024-12-31T23:59:59Z",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (for pagination)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (for pagination)",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all check-in logs (optional filter by member, date range, status)",
        "tags": ["Admin - Check-ins"]
      }
    },
    "/api/v1/members/{id}/checkins": {
      "get": {
        "operationId": "CheckInController_findByMemberId_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": false,
            "in": "query",
            "description": "Filter by member ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by check-in status",
            "schema": {
              "enum": ["ALLOWED", "DENIED"],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for date range filter (ISO 8601)",
            "schema": {
              "example": "2024-01-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for date range filter (ISO 8601)",
            "schema": {
              "example": "2024-12-31T23:59:59Z",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (for pagination)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (for pagination)",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List check-in logs for a specific member",
        "tags": ["Admin - Check-ins"]
      }
    },
...
  },
  "info": {
    "title": "Gym Mini Check-in System API",
    "description": "API documentation for Gym Mini Check-in System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}